+++ src/main/java/org/spongepowered/api/item/inventory/ItemStackSnapshot.java
@@ -24,11 +24,8 @@
  */
 package org.spongepowered.api.item.inventory;
 
-import net.kyori.adventure.text.event.HoverEvent;
-import net.kyori.adventure.text.event.HoverEventSource;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.data.SerializableDataHolder;
-import org.spongepowered.api.item.ItemType;
 
 /**
  * Represents a snapshot of an {@link ItemStack} as an
@@ -37,7 +34,7 @@ import org.spongepowered.api.item.ItemType;
  * it is a snapshot, a snapshot cannot be modified, but modifications will
  * result in a new instance of the {@link ItemStackSnapshot}.
  */
-public interface ItemStackSnapshot extends HoverEventSource<HoverEvent.ShowItem>, SerializableDataHolder.Immutable<ItemStackSnapshot> {
+public interface ItemStackSnapshot extends ItemStackLike, SerializableDataHolder.Immutable<ItemStackSnapshot> {
 
     /**
      * Gets a empty {@link ItemStackSnapshot}.
@@ -47,38 +44,6 @@ public interface ItemStackSnapshot extends HoverEventSource<HoverEvent.ShowItem>
     static ItemStackSnapshot empty() {
         return Sponge.game().factoryProvider().provide(Factory.class).empty();
     }
-
-    /**
-     * Gets the {@link ItemType} of this {@link ItemStackSnapshot}. The
-     * {@link ItemType} is always available.
-     *
-     * @return The item type
-     */
-    ItemType type();
-
-    /**
-     * Gets the quantity of items in this the {@link ItemStack} this
-     * {@link ItemStackSnapshot} is representing.
-     *
-     * @return The current stack size
-     */
-    int quantity();
-
-    /**
-     * Returns true if {@link #quantity()} is zero and therefore this
-     * ItemStackSnapshot is empty.
-     *
-     * @return True if this ItemStackSnapshot is empty
-     */
-    boolean isEmpty();
-
-    /**
-     * Creates a new {@link ItemStack} with all the data currently available
-     * on this {@link ItemStackSnapshot}.
-     *
-     * @return The newly generated item stack
-     */
-    ItemStack createStack();
 
     interface Factory {
